remove references to opensiddur-tests module
