+++ lib/restxq.registry
@@ -2,7 +2,6 @@ version: 1
 /db/apps/opensiddur-demos/api/demo.xqm,{http://jewishliturgy.org/api/demo}list#0{http://jewishliturgy.org/api/demo}transliterate-xml#2{http://jewishliturgy.org/api/demo}transliterate-text#2{http://jewishliturgy.org/api/demo}transliteration-list#3
 /db/apps/opensiddur-server/api/data/conditionals.xqm,{http://jewishliturgy.org/api/data/conditionals}post#1{http://jewishliturgy.org/api/data/conditionals}delete#1{http://jewishliturgy.org/api/data/conditionals}list#5{http://jewishliturgy.org/api/data/conditionals}get#1{http://jewishliturgy.org/api/data/conditionals}put#2
 /db/apps/opensiddur-server/api/data/styles.xqm,{http://jewishliturgy.org/api/data/styles}get-access#2{http://jewishliturgy.org/api/data/styles}list#3{http://jewishliturgy.org/api/data/styles}get-css#1{http://jewishliturgy.org/api/data/styles}post#1{http://jewishliturgy.org/api/data/styles}delete#1{http://jewishliturgy.org/api/data/styles}put-access#2{http://jewishliturgy.org/api/data/styles}put-xml#2{http://jewishliturgy.org/api/data/styles}get-xml#1{http://jewishliturgy.org/api/data/styles}put-css#2
-/db/apps/opensiddur-tests/api/tests.xqm,{http://jewishliturgy.org/api/tests}get#2
 /db/apps/opensiddur-server/api/data/original.xqm,{http://jewishliturgy.org/api/data/original}get-combined#2{http://jewishliturgy.org/api/data/original}delete#1{http://jewishliturgy.org/api/data/original}put-access#2{http://jewishliturgy.org/api/data/original}post#1{http://jewishliturgy.org/api/data/original}post-combined-job#3{http://jewishliturgy.org/api/data/original}get-flat#1{http://jewishliturgy.org/api/data/original}list#3{http://jewishliturgy.org/api/data/original}put-flat#2{http://jewishliturgy.org/api/data/original}get#1{http://jewishliturgy.org/api/data/original}put#2{http://jewishliturgy.org/api/data/original}get-combined-html-forced#2{http://jewishliturgy.org/api/data/original}get-access#2{http://jewishliturgy.org/api/data/original}get-combined-html#2
 /db/apps/opensiddur-server/api/group.xqm,{http://jewishliturgy.org/api/group}put#2{http://jewishliturgy.org/api/group}get-user-groups#1{http://jewishliturgy.org/api/group}list#2{http://jewishliturgy.org/api/group}delete#1{http://jewishliturgy.org/api/group}get-html#1{http://jewishliturgy.org/api/group}get-xml#1
 /db/apps/opensiddur-server/api/data/linkage.xqm,{http://jewishliturgy.org/api/data/linkage}get-access#2{http://jewishliturgy.org/api/data/linkage}list#3{http://jewishliturgy.org/api/data/linkage}get#1{http://jewishliturgy.org/api/data/linkage}put#2{http://jewishliturgy.org/api/data/linkage}delete#1{http://jewishliturgy.org/api/data/linkage}put-access#2{http://jewishliturgy.org/api/data/linkage}post#1
