+++ opensiddur-server/src/api/index.xqm
@@ -70,15 +70,6 @@ declare
             <a class="discovery" href="{$api-base}/static">Static files</a>
           </li>,
           <li class="api">
-            <a class="discovery" href="{$api-base}/test">Test (XQSuite)</a>
-          </li>,
-          index:if-installed(
-            "http://jewishliturgy.org/apps/opensiddur-tests",
-            <li class="api">
-              <a class="discovery" href="{$api-base}/tests">Tests</a>
-            </li>
-          ),
-          <li class="api">
             <a class="discovery" href="{$api-base}/user">User</a>
           </li>,
           <li class="api">
