--- opensiddur-tests/build.xml
@@ -1,54 +1,0 @@
-<project name="opensiddur-tests" default="dist" basedir=".">
-	<description>
-        Build file for opensiddur-tests
-    </description>
-	<!-- set global properties for this build -->
-	<dirname property="base.dir" file="${ant.file.opensiddur-tests}"/>
-    <property name="lib.dir" location="../lib"/>
-	<property name="src" location="${base.dir}/src"/>
-    <property name="build" location="${base.dir}/build"/>
-	<property name="dist"  location="${base.dir}/dist"/>
-
-	<!-- set local properties for the build -->
-	<property file="local.build.properties"/>
-    <!-- set global default properties for the build -->
-    <property file="build.properties"/>
-	
-	<target name="init">
-		<!-- Create the time stamp -->
-		<tstamp/>
-        <!-- make build dirs -->
-        <mkdir dir="${build}"/>
-        <mkdir dir="${dist}"/>
-	</target>
-
-	<target name="dist" depends="init,xar"
-		description="build everything"/>
-
-	<target name="xar"
-        depends="init" 
-        description="generate the xar distribution" >
-        <copy todir="${build}">
-			<fileset dir="${src}">
-				<include name="**/*.xml"/>
-				<include name="**/*.xq?"/>
-				<include name="**/*.html"/>
-				<include name="**/*.js"/>
-				<exclude name="**/*.xar" />
-				<exclude name="build.xml"/>
-			</fileset>
-        </copy>
-		<zip 
-            destfile="${dist}/opensiddur-tests.xar" 
-            basedir="${build}">
-		</zip>
-	</target>
-
-	<target name="clean"
-        description="clean up">
-		<!-- Delete the ${build} and ${dist} directory trees -->
-	    <delete dir="${build}"/>
-    	<delete dir="${dist}"/>
-	</target>
-</project>
-
