--- opensiddur-tests/src/expath-pkg.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<package 
-  xmlns="http://expath.org/ns/pkg" 
-  name="http://jewishliturgy.org/apps/opensiddur-tests" 
-  abbrev="opensiddur-tests" 
-  version="0.12.0"
-  spec="1.0">
-    <title>Open Siddur Tests</title>
-    <dependency package="http://jewishliturgy.org/apps/opensiddur-server" semver="0.12.0"/>
-</package>
