--- opensiddur-tests/src/index.html
@@ -1,148 +1,0 @@
-<!DOCTYPE html>
-<html 
-  lang="en" 
-  xml:lang="en" 
-  xmlns="http://www.w3.org/1999/xhtml"
-  data-ng-app="OpenSiddurTestingApp"
-  >
-  <head>
-    <meta charset="utf-8"/>
-    <meta http-equiv="Content-Type" content="application/xhtml+xml; charset=UTF-8"/>
-    <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
-    <meta name="description" content="Open Siddur web application"/>
-    <meta name="author" content="Efraim Feinstein"/>
-    <title>Open Siddur Testing Application</title>
-    <!-- bootstrap boilerplate -->
-    <link href="http://netdna.bootstrapcdn.com/twitter-bootstrap/2.3.2/css/bootstrap-combined.min.css" rel="stylesheet"/>
-    <style type="text/css">
-      body {
-        padding-top: 60px;
-        padding-bottom: 40px;
-      }
-      .sidebar-nav {
-        padding: 9px 0;
-      }
-      
-      .pass {
-          background-color:green;
-      }
-      
-      .fail {
-          background-color:red;
-      }
-      
-      .ignore {
-          background-color:darkgrey;
-      }
-
-      @media (max-width: 980px) {
-        /* Enable use of floated navbar text */
-        .navbar-text.pull-right {
-          float: none;
-          padding-left: 5px;
-          padding-right: 5px;
-        }
-      }
-    </style>
-  </head>
-  <body>
-    <!-- top nav bar -->
-    <div class="navbar navbar-inverse navbar-fixed-top">
-      <div class="navbar-inner">
-        <div class="container-fluid">
-          <a class="brand" href="#">Open Siddur Project: Pre-release testing</a>
-        </div>
-      </div>
-    </div>
-    <div data-ng-controller="MainCtrl">
-      <div class="row-fluid">
-        <button class="btn btn-large btn-primary" type="submit" data-ng-click="runAll()"  data-ng-disabled="runningAll">{{runningButtonText()}}</button>
-      </div>
-      <div class="row-fluid">
-        <div class="span12">
-          <table border="1" class="summary">
-            <tr>
-              <th>Tests run</th>
-              <th>Passed</th>
-              <th>Failed</th>
-              <th>Ignored</th>
-            </tr>
-            <tr>
-                <td class="total">{{passed + failed + ignored}}</td>
-                <td data-ng-class="{pass: passed > 0}">{{passed}}</td>
-                <td data-ng-class="{fail: failed > 0}">{{failed}}</td>
-                <td data-ng-class="{ignore: ignored > 0}">{{ignored}}</td>
-            </tr>  
-          </table>  
-          <div data-ng-repeat="suite in suites">
-            <h2>
-            <button data-ng-hide="suite.running || runningAll" 
-                data-ng-click="runSuite(suite.n)"
-                class="btn">Run</button> 
-            {{suite.name}} |
-                <span data-ng-class="{pass: suite.passed > 0}">Pass: {{suite.passed}}</span>| 
-                <span data-ng-class="{fail: suite.failed > 0}">Fail: {{suite.failed}}</span>|
-                <span data-ng-class="{ignore: suite.ignored > 0}">Ignore: {{suite.ignored}}</span>
-            </h2>
-            <div data-ng-show="suite.running">Running tests...</div>
-            <div data-ng-show="suite.complete">
-              <table border="1" class="summary">
-                <tr>
-                  <th>Tests</th>
-                  <th>Passed</th>
-                  <th>Failed</th>
-                  <th>Ignored</th>
-                </tr>
-                <tr>
-                    <td class="total">{{suite.total}}</td>
-                    <td data-ng-class="{pass: suite.passed > 0}">{{suite.passed}}</td>
-                    <td data-ng-class="{fail: suite.failed > 0}">{{suite.failed}}</td>
-                    <td data-ng-class="{ignore: suite.ignored > 0}">{{suite.ignored}}</td>
-                </tr>  
-              </table>  
-              <a data-ng-hide="suite.showDetails" data-ng-click="suite.toggleDetails()">Show details &gt;&gt; </a>
-              <div class="span12" data-ng-show="suite.showDetails">
-                <a data-ng-click="suite.toggleDetails()">&lt;&lt; Hide details</a>
-                <h3>{{suite.details.description}}</h3>
-                <div class="span10" data-ng-repeat="testSet in suite.details.testSets">
-                    <h4>{{testSet.description}} |
-                        <span data-ng-class="{pass: testSet.passed > 0}">Pass: {{testSet.passed}}</span>| 
-                        <span data-ng-class="{fail: testSet.failed > 0}">Fail: {{testSet.failed}}</span>|
-                        <span data-ng-class="{ignore: testSet.ignored > 0}">Ignore: {{testSet.ignored}}</span>
-                    </h4>
-                    <div data-ng-repeat="test in testSet.tests">
-                        <h5>{{test.description}} | 
-                          <span data-ng-class="{pass: test.passed > 0}">Pass: {{test.passed}}</span>| 
-                          <span data-ng-class="{fail: test.failed > 0}">Fail: {{test.failed}}</span>|
-                          <span data-ng-class="{ignore: test.ignored > 0}">Ignore: {{test.ignored}}</span>
-                        </h5>
-                        <div data-ng-repeat="assertion in test.assertions">
-                            <h6 class="{{assertion.pass}}">{{assertion.description}}: {{assertion.pass}}</h6>
-                            <div data-ng-show="{{assertion.pass == 'fail'}}"><pre>{{assertion.error}}</pre></div>
-                        </div>
-                    </div>
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-    <div id="footer" class="container-fluid">
-      <hr/>
-  
-      <footer>
-        <p>&#169; 2013 <a href="http://opensiddur.org">Contributors to the Open Siddur Project</a>. See <a href="https://github.com/opensiddur">our github repository</a> for the source code and information about open source licensing.</p>
-      </footer>  
-    </div><!--/.fluid-container-->
-    
-    <!-- Le javascript
-    ================================================== -->
-    <!-- Placed at the end of the document so the pages load faster -->
-    <script src="http://code.jquery.com/jquery-2.0.3.min.js" ></script>
-    <script src="http://code.angularjs.org/1.1.5/angular.min.js" ></script>
-    <script src="http://netdna.bootstrapcdn.com/twitter-bootstrap/2.3.2/js/bootstrap.min.js" ></script>
-    <script src="js/app.js" ></script>
-  
-  </body>
-</html>
