--- opensiddur-tests/src/pre-install.xql
@@ -1,4 +1,0 @@
-xquery version "3.0";
-(:~ pre-install script for tests :)
-
-util:log-system-out("tests pre-install disabled due to opensiddur issue #156")
