--- opensiddur-tests/src/tests/api/template.t.tmpl.xml
@@ -1,35 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<TestSuite>
-  <suiteName>NAME ME</suiteName>
-  <description>
-      <p>Test SOMETHING</p>
-      <author>Efraim Feinstein</author>
-      <copyright>Copyright 2012 Efraim Feinstein,
-    Licensed under the GNU Lesser General Public License, version 3 or later</copyright>
-  </description>
-  <setup/>
-  <tearDown/>
-  <namespace prefix="html">http://www.w3.org/1999/xhtml</namespace>
-  <namespace prefix="tei">http://www.tei-c.org/ns/1.0</namespace>
-  <namespace prefix="j">http://jewishliturgy.org/ns/jlptei/1.0</namespace>
-  <namespace prefix="jx">http://jewishliturgy.org/ns/jlp-processor</namespace>
-  <imports>
-    import module namespace {NS}="http://jewishliturgy.org/{MODULE}"
-      at "xmldb:exist:///code/{PATH}.xqm";
-  </imports>
-  <TestSet>
-      <testName>{NAME}</testName>
-      <setup/>
-      <imports>
-    </imports>
-      <functions>
-    </functions>
-      <tearDown/>
-      <test>
-         <task>PUT</task>
-         <code>
-        apitest:delete("***", ())
-      </code> 
-      </test>
-  </TestSet>
-</TestSuite>
