--- opensiddur-tests/src/tests/contributors.xml
@@ -1,87 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<tei:TEI xmlns:tei="http://www.tei-c.org/ns/1.0"
-         xml:base="http://jewishliturgy.org/base/text/contributors">
-    <tei:teiHeader>
-        <tei:fileDesc>
-            <tei:titleStmt>
-                <tei:title xml:lang="en">Fake contributor list for database unit tests</tei:title>
-            </tei:titleStmt>
-            <tei:publicationStmt>
-                <tei:availability>
-                    <tei:p xml:lang="en">To the extent possible under law, 
-     the contributors who associated  
-     <tei:ref type="license" target="licenses/cc0/1.0">Creative Commons Zero</tei:ref> 
-     with this work have waived all copyright and related or
-     neighboring rights to this work.  
-     A list of contributors is available at 
-     <tei:ref type="attribution" target="http://jewishliturgy.org/base/text/contributors">http://jewishliturgy.org/base/text/contributors</tei:ref>.
-     </tei:p>
-                    <rdf:RDF xmlns:rdf="http://www.w3.org/1999/02/22-rdf-syntax-ns#">
-                        <cc:License xmlns:cc="http://web.resource.org/cc/"
-                              rdf:about="http://creativecommons.org/publicdomain/zero/1.0/">
-                            <cc:legalcode rdf:resource="http://creativecommons.org/publicdomain/zero/1.0/legalcode"/>
-                            <cc:permits rdf:resource="http://creativecommons.org/ns#Reproduction"/>
-                            <cc:permits rdf:resource="http://creativecommons.org/ns#Distribution"/>
-                        </cc:License>
-                    </rdf:RDF>
-                </tei:availability>
-                <tei:idno type="svn">$Id: contributors.xml 411 2010-01-03 06:58:09Z efraim.feinstein $</tei:idno>
-            </tei:publicationStmt>
-            <tei:sourceDesc>
-                <!--  this comment is used in testing numerical-path() -->
-                <tei:p>Born digital.</tei:p>
-                <!--  this comment is also used in testing numerical-path() -->
-            </tei:sourceDesc>
-        </tei:fileDesc>
-      <!-- if you have an application, insert the information here:
-  <tei:encodingDesc>
-   <tei:appInfo>
-    <tei:application ident="***" version="***">
-     <tei:label>***</tei:label>
-     <tei:p>***Description***</tei:p>
-    </tei:application>
-   </tei:appInfo>
-  </tei:encodingDesc>
-  -->
-    </tei:teiHeader>
-    <tei:text>
-        <tei:body>
-            <tei:div type="contributors">
-                <tei:list>
-                    <tei:item xml:id="efraim.feinstein">
-                        <tei:forename>Efraim</tei:forename>
-                        <tei:surname>Feinstein</tei:surname>, 
-        <tei:email>efraim.feinstein@gmail.com</tei:email>
-                    </tei:item>
-                    <tei:item xml:id="azriel.fasten">
-                        <tei:forename>Azriel</tei:forename>
-                        <tei:surname>Fasten</tei:surname>
-                        <tei:email>fst911@gmail.com</tei:email>
-                    </tei:item>
-                    <tei:item xml:id="test">
-                        <tei:forename>Mister</tei:forename>
-                        <tei:forename>Samuel</tei:forename>
-                        <tei:surname>Test</tei:surname>
-                        <tei:email>test@example.com</tei:email>
-                        <tei:ptr type="openid" target="http://test3.blogspot.com"/>
-                    </tei:item>
-										     <tei:item xml:id="toby.replaced">
-												      <tei:forename>Toby</tei:forename>
-												      <tei:surname>Replaced</tei:surname>
-												      <tei:email>toby@replac.edu</tei:email>
-										     </tei:item>
-                    <tei:item xml:id="toby.deleted">
-                      <tei:forename>Toby</tei:forename>
-                      <tei:surname>Deleted</tei:surname>
-                      <tei:email>toby@delet.edu</tei:email>
-                    </tei:item>
-                    <tei:item xml:id="toby.prepared">
-                      <tei:forename>Toby</tei:forename>
-                      <tei:surname>Peppered</tei:surname>
-                      <tei:email>toby@prepar.edu</tei:email>
-                    </tei:item>
-                </tei:list>
-            </tei:div>
-        </tei:body>
-    </tei:text>
-</tei:TEI>
