--- opensiddur-tests/src/tests/minimal-invalid.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<tei:TEI xmlns:tei="http://www.tei-c.org/ns/1.0"
-         xml:base="http://jewishliturgy.org/base/text/contributors">
-  <!-- missing header parts -->
-    <tei:teiHeader>
-      <tei:publicationStmt>
-        <tei:availability>
-          <!-- this is here to set off the header Schematron -->
-        </tei:availability>
-      </tei:publicationStmt>
-    </tei:teiHeader>
-    <tei:text>
-        <tei:body>
-            <tei:div xml:lang="en">
-              <tei:head>Shortest document ever</tei:head>
-              <tei:p>Yay!</tei:p>
-            </tei:div>
-        </tei:body>
-    </tei:text>
-</tei:TEI>
