--- opensiddur-tests/src/tests/minimal-valid.xml
@@ -1,43 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<tei:TEI xmlns:tei="http://www.tei-c.org/ns/1.0"
-         xml:base="http://jewishliturgy.org/base/text/contributors">
-    <tei:teiHeader>
-        <tei:fileDesc>
-            <tei:titleStmt>
-                <tei:title xml:lang="en">Minimal valid JLPTEI</tei:title>
-            </tei:titleStmt>
-            <tei:publicationStmt>
-                <tei:availability>
-                    <tei:p xml:lang="en">To the extent possible under law, 
-     the contributors who associated  
-     <tei:ref type="license" target="licenses/cc0/1.0">Creative Commons Zero</tei:ref> 
-     with this work have waived all copyright and related or
-     neighboring rights to this work.  
-     A list of contributors is available at 
-     <tei:ref type="attribution" target="http://jewishliturgy.org/base/text/contributors">http://jewishliturgy.org/base/text/contributors</tei:ref>.
-     </tei:p>
-                    <rdf:RDF xmlns:rdf="http://www.w3.org/1999/02/22-rdf-syntax-ns#">
-                        <cc:License xmlns:cc="http://web.resource.org/cc/"
-                              rdf:about="http://creativecommons.org/publicdomain/zero/1.0/">
-                            <cc:legalcode rdf:resource="http://creativecommons.org/publicdomain/zero/1.0/legalcode"/>
-                            <cc:permits rdf:resource="http://creativecommons.org/ns#Reproduction"/>
-                            <cc:permits rdf:resource="http://creativecommons.org/ns#Distribution"/>
-                        </cc:License>
-                    </rdf:RDF>
-                </tei:availability>
-                <tei:idno type="svn">$Id: minimal-valid.xml 411 2010-01-03 06:58:09Z efraim.feinstein $</tei:idno>
-            </tei:publicationStmt>
-            <tei:sourceDesc>
-                <tei:p>Born digital.</tei:p>
-            </tei:sourceDesc>
-        </tei:fileDesc>
-    </tei:teiHeader>
-    <tei:text>
-        <tei:body>
-            <tei:div xml:lang="en">
-              <tei:head>Shortest document ever</tei:head>
-              <tei:p>Yay!</tei:p>
-            </tei:div>
-        </tei:body>
-    </tei:text>
-</tei:TEI>
