--- opensiddur-tests/src/tests/transforms/flatten/no-concurrency.xml
@@ -1,28 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<TEI xmlns="http://www.tei-c.org/ns/1.0" xml:lang="en">
-  <teiHeader>
-      <fileDesc>
-         <titleStmt>
-            <title type="main">Null</title>
-         </titleStmt>
-         <publicationStmt>
-            <availability>
-               <licence target="http://creativecommons.org/publicdomain/zero/1.0"/>
-            </availability>
-            <distributor>Open Siddur Project</distributor>
-            <date when="2012">2012-06-27</date>
-         </publicationStmt>
-         <sourceDesc>
-            <tei:bibl xmlns:tei="http://www.tei-c.org/ns/1.0" xmlns="">
-               <tei:ptr type="bibl" target="/data/sources/open_siddur_project"/>
-               <tei:ptr type="bibl-content" target="#text"/>
-            </tei:bibl>
-         </sourceDesc>
-      </fileDesc>
-  </teiHeader>
-  <text>
-      <j:streamText xmlns:j="http://jewishliturgy.org/ns/jlptei/1.0" xml:id="text">
-         <seg xml:id="seg1">Segment!</seg>
-      </j:streamText>
-  </text>
-</TEI>
