--- opensiddur-tests/src/tests/transforms/flatten/resolve1.xml
@@ -1,40 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<tei:TEI xmlns:tei="http://www.tei-c.org/ns/1.0" xml:lang="en">
-  <tei:teiHeader>
-    <!-- leaving it blank... -->
-  </tei:teiHeader>  
-  <tei:text>
-      <j:concurrent xmlns:j="http://jewishliturgy.org/ns/jlptei/1.0">
-         <jf:merged xmlns:jf="http://jewishliturgy.org/ns/jlptei/flat/1.0">
-            <tei:div jf:start="div1" jf:position="1" jf:relative="-1" jf:nchildren="-1"
-                     jf:nlevels="1"
-                     jf:nprecedents="1"
-                     jf:layer-id="div-layer"
-                     jf:id="div1"/>
-            <tei:head jf:position="1" jf:relative="-1" jf:nchildren="-1" jf:nlevels="2"
-                      jf:nprecedents="1"
-                      jf:layer-id="div-layer">title</tei:head>
-            <tei:ab jf:start="..." jf:position="1" jf:relative="-1" jf:nchildren="-1"
-                    jf:nlevels="2"
-                    jf:nprecedents="2"
-                    jf:layer-id="div-layer"/>
-            <tei:label jf:position="1" jf:relative="-1" jf:nchildren="-1" jf:nlevels="3"
-                       jf:nprecedents="1"
-                       jf:layer-id="div-layer">Label</tei:label>
-            <jf:placeholder jf:position="1" jf:relative="0" jf:nchildren="0" jf:nlevels="0"
-                            jf:nprecedents="0"
-                            jf:stream="stream"
-                            jf:id="seg1"/>
-            <tei:ab jf:end="..." jf:position="1" jf:relative="1" jf:nchildren="1" jf:nlevels="-2"
-                    jf:nprecedents="2"
-                    jf:layer-id="div-layer"/>
-            <tei:div jf:end="div1" jf:position="1" jf:relative="1" jf:nchildren="1" jf:nlevels="-1"
-                     jf:nprecedents="1"
-                     jf:layer-id="div-layer"/>
-         </jf:merged>
-      </j:concurrent>
-      <j:streamText xmlns:j="http://jewishliturgy.org/ns/jlptei/1.0" xml:id="stream">
-         <tei:seg xml:id="seg1">One segment</tei:seg>
-      </j:streamText>
-  </tei:text>
-</tei:TEI>
